General: Code cleanup
* Restore the config macros. Now hopefully fixed with the
  Zc:preprocessor compiler option

* Use CMAKE_MSVC_RUNTIME_LIBRARY instead of replacing flags
  manually. (They didn't work before because minimum cmake level
  wasn't high enough and as such the policy that enabled the option
  wasn't active)

* Renamed rainbomizer.asi output to use the project
  name (V.Rainbomizer)

* Added Forced Vehicle config option

* Removed unused natives pattern searches

+++ lib/rage.hh
@@ -326,15 +326,15 @@ struct atFixedArray : public atArrayBase<T[Size]>
 struct atString
 {
     char *m_szString;
-    short m_nLength;
-    short m_nCapacity;
+    unsigned short m_nLength;
+    unsigned short m_nCapacity;
 
     atString () = default;
 
     atString (const char *str)
     {
         m_szString  = const_cast<char *> (str);
-        m_nLength   = strlen (str);
+        m_nLength   = static_cast<unsigned short> (strlen (str));
         m_nCapacity = m_nLength;
     }
 };
